data_analysis_agents/requirements.txt
# Core LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0

# Telemetry
langsmith>=0.1.0
langfuse>=2.0.0

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Data Analysis
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
scipy>=1.12.0
openpyxl>=3.1.0

# Code Safety
RestrictedPython>=7.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
